refactor(observability,quarantine): reduce cyclomatic complexity
Split large modules into smaller, focused modules:

Anomaly detection:
- anomaly.py (445 LOC) -> anomaly/ package:
  - types.py: Base types (Anomaly, enums)
  - detector.py: Main AnomalyDetector class
  - monitor.py: DataQualityMonitor
  - detectors/: Strategy pattern implementations
    - base.py: DetectorStrategy protocol
    - zscore.py: Z-score detection
    - iqr.py: IQR-based detection
    - mad.py: MAD-based detection

Quarantine:
- unified_quarantine.py (495 LOC) -> split into:
  - unified.py: Main UnifiedQuarantine class (176 LOC)
  - operations.py: Read operations (168 LOC)
  - helpers.py: Utility functions (42 LOC)
  - unified_quarantine.py: Backward compatibility shim

Results:
- All files < 200 LOC (max: 176 lines)
- xenon --max-absolute B passes (CC ≤ 10)
- All 80 tests pass
- Full backward compatibility maintained
