+++ src/bioetl/infrastructure/observability/anomaly/detectors/base.py
@@ -1,0 +1,59 @@
+"""Base protocol for anomaly detection strategies.
+
+Defines interface that all detection algorithms must implement.
+"""
+
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
+
+    from bioetl.infrastructure.observability.anomaly.types import (
+        Anomaly,
+        AnomalySeverity,
+    )
+
+
+class DetectorStrategy(ABC):
+    """Abstract base for anomaly detection algorithms.
+
+    All detection strategies must implement:
+    - detect(): Analyze value against baseline
+    - get_severity(): Map score to severity level
+
+    Strategies should be stateless; baseline data is passed to detect().
+    """
+
+    @abstractmethod
+    def detect(
+        self,
+        metric_name: str,
+        current_value: float,
+        baseline: Sequence[float],
+        threshold: float,
+    ) -> Anomaly | None:
+        """Detect anomaly in current value.
+
+        Args:
+            metric_name: Name of metric being analyzed
+            current_value: Current observed value
+            baseline: Historical values for comparison
+            threshold: Detection threshold (interpretation varies by strategy)
+
+        Returns:
+            Anomaly if detected, None otherwise
+        """
+
+    @abstractmethod
+    def get_severity(self, score: float) -> AnomalySeverity:
+        """Map detection score to severity level.
+
+        Args:
+            score: Detection score (e.g., z-score, IQR multiplier)
+
+        Returns:
+            Severity level based on score
+        """
