+++ src/bioetl/infrastructure/quarantine/__init__.py
@@ -2,3 +2,7 @@
 
 Implements RULES.md §2.6 - Quarantine Policy.
 """
+
+from bioetl.infrastructure.quarantine.unified import UnifiedQuarantine
+
+__all__ = ["UnifiedQuarantine"]
